chore: rename files, remove obsolete
- remove ingredient slice, active ingredient is not used
- rename initialSlice's, update usage
- remove logs
