+++ src/services/burger-constructor/index.js
@@ -1,5 +1,5 @@
 import { createSlice } from "@reduxjs/toolkit";
-import { initialState } from "./initialState";
+import { initialState } from "./initial-state";
 import * as reducers from "./reducers";
 import * as selectors from "./selectors";
