+++ src/services/burger-constructor/reducers/clear-burger.js
@@ -1,3 +1,3 @@
-import { initialState } from "../initialState";
+import { initialState } from "../initial-state";
 
 export const clearBurger = () => initialState;
