--- src/services/ingredient/initialState.js
@@ -1,3 +1,0 @@
-export const initialState = {
-  ingredient: null,
-};
