--- src/services/protected-route/initial-state.js
@@ -1,3 +1,0 @@
-export const initialState = {
-  isAuthChecked: false,
-};
