--- src/services/protected-route/thunks/check-token.js
@@ -1,7 +1,0 @@
-import { createAsyncThunk } from "@reduxjs/toolkit";
-import { burgersApiController } from "../../../utils/api/burgers-api-controller";
-
-export const checkToken = createAsyncThunk(
-  "protectedRoute/updateToken",
-  burgersApiController.checkToken,
-);
