+++ src/services/root-reducer.js
@@ -1,7 +1,6 @@
 import { combineReducers } from "@reduxjs/toolkit";
 import { ingredientsSlice } from "./ingredients";
 import { orderSlice } from "./order";
-import { ingredientSlice } from "./ingredient";
 import { burgerConstructorSlice } from "./burger-constructor";
 import { ingredientsTabsSlice } from "./ingredients-tabs";
 import {
@@ -16,7 +15,6 @@ import { profileSlice } from "./profile";
 const appReducer = combineReducers({
   [ingredientsSlice.reducerPath]: ingredientsSlice.reducer,
   [orderSlice.reducerPath]: orderSlice.reducer,
-  [ingredientSlice.reducerPath]: ingredientSlice.reducer,
   [burgerConstructorSlice.reducerPath]: burgerConstructorSlice.reducer,
   [ingredientsTabsSlice.reducerPath]: ingredientsTabsSlice.reducer,
   [formRegisterSlice.reducerPath]: formRegisterSlice.reducer,
